requirements.txt
# OpenClaw Requirements

# Data fetching
tushare>=1.3.0
pytdx>=1.4.0
yfinance>=0.2.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Environment
python-dotenv>=1.0.0

# HTTP requests
requests>=2.28.0

# Technical analysis
ta-lib>=0.4.28

# Time series forecasting
timesfm>=1.0.0

# Web search
tavily>=0.1.0

# LLM / AI
anthropic>=0.18.0
google-generativeai>=0.3.0

# Database
sqlite3 (built-in)

# PDF processing
pymupdf>=1.23.0
pdfplumber>=0.10.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Image processing
pillow>=10.0.0
openai>=1.0.0

# Vector database
chromadb>=0.4.0

# Jupyter
jupyter>=1.0.0
ipython>=8.0.0
